feat(api): Add support for get initialized status
The endpoint GET /sys/init was already defined but not exposed
for use. An API getInitStatus is now available for use to verify
if the vault has been initialized or not.

Tests for each of the init APIs have been added to validate the
functionality. Testing will now require a functional instance
of vault running. To test make sure the vault binary is available
on the local path and run `vault server -config=config/dev.hcl`