chore(tests): Test and document core
Implement test cases for the core library and documentation of each
of the public methods.

As a result of adding test cases for all modules, all API methods
will now return a promise that will be resolved or rejected. And
several bug fixes or inconsistencies were resolved.

The init method was made more user-friendly such that if the Vault
has already been initialized the Vault is returned instead of rejecting
the request.

Because the internal state of the Vault of critical and the ability to
repeatedly restart the Vault is required, internal state management has
been made to be recoverable.