fix(apis): Token per API call - chained
If the API calls are chained and a token was passed in, the token
is deleted so the additional calls in the chain fail.
For internal use of chaining, make sure token is passed along.
Stop deleting from token from options. No longer maintains a copy
of the policies as part of internal state as it provides no value.