fix(seal): Require master key share to unseal
unseal requires master key share as input

**BREAKING CHANGE:** Drops the capturing keys and token automatically
on init and now a master key share must be passed in. The orchestration
of calling unseal with different master key shares is now the
responsibility of the consumer just like Vault.

Also removes backing up of the token and keys as the state is no longer
held internally and no longer captured as part of `init`. The caller is
the responsible to capture information and manage directly.

Refer to `getting_started.md` for how to do `init` and `unseal`.

closes #65