fix(init): initialized not being set
When passing the auth token to each API and not using `setToken` the
`initialized` flag is not so all APIs will return
`Vault has not been initialized` error.

As part of performing executing `init` or `prepare` will make sure the
`initialized` flag has been properly set. And test cases now included to
verify this functionality.

closes #70